requirements.txt
requests
Flask
Flask-Cors
flask-restful
psycopg2-binary
python-dotenv